Fixed precision of most database fields.

+++ db/init/001_init.sql
@@ -7,20 +7,20 @@ ALTER DATABASE gle OWNER TO gle;
 
 CREATE TABLE batteries (
 	id SERIAL,
-	power REAL DEFAULT 0,
-	max_capacity REAL,
+	power NUMERIC(5, 2) DEFAULT 0,
+	max_capacity NUMERIC(5, 0),
 	PRIMARY KEY (id)
 );
 ALTER TABLE batteries OWNER TO gle;
 
 CREATE TABLE prosumers (
 	id SERIAL,
-	mean_day_wind_speed REAL,
-	current_wind_speed REAL,
-	current_consumption REAL,
-	current_production REAL,
-	ratio_excess_market REAL,
-	ratio_deficit_market REAL,
+	mean_day_wind_speed NUMERIC(5, 2),
+	current_wind_speed NUMERIC(5, 2),
+	current_consumption NUMERIC(5, 2),
+	current_production NUMERIC(5, 2),
+	ratio_excess_market NUMERIC(2, 1),
+	ratio_deficit_market NUMERIC(2, 1),
 	battery_id INTEGER,
 	blackout BOOLEAN DEFAULT false,
 	PRIMARY KEY (id),
